Create README.md
IVELT App for Android Devices (maybee soon windows 11 ;)) (For flipphones as well)

Ivelt app that can not access any nonkosher websites. Unfortunately, it's not on the playstore, and there are no plans to launch it there either. If you'd like to see it on playstore, please let me know in the issues tab (any support is also graciously accepted).

Official Releases can be seen at https://github.com/KFMDMSolutions/Ivelt/releases.


So what does this app do?
Ivelt app is intended for users running Android Smart/Flipphone (Min SDK 16)


FAQs

So, how does it work? You have the exact code in front of you, feel free to read it { It's not that bad ;) }. In a nutshell, this app uses Web app features provided by Android.

I think there's something that can be improved. How to contribute? We welcome all kinds of contributions. Documentation, Code tests, Actual Hardware Testing etc etc. Please feel free to create a Pull request and remind me in a couple of weeks if I forget about it :D

Is there a version for Flip Phones too? Yes! this app also works on flip phones as it is named with the android.cts.example package name.

Are there ADs? No, there are no ads, and there won't be any from our end but IVELT does have ads so the ads that are on the site will still come up, but by clicking it you wont be able to access any other sites.

Is there any kind of tracking information? Do you sell it for profit? No, there is no tracking information collected from the app. No telemetry nothing.

How do you sustain the project? It's currently on Ad-Hoc basis. This project is not funded by anyone. It totally relies on community support.


Redistributing and Creating MODs
We welcome everyone who'd like to work on the application, but I request everyone to stick to the following

Please consider sending the enhancement to this project itself. Were open to all enhancements and fixes. The whole point of making this app open source is so that people have full access to the source code.
Please considering pointing to the latest releases from this repository itself, unless you explicitly want folks to download an old version. This prevents more people from downloading old versions and reporting bugs that might have been fixed.
If you want to make a commercial application out of the project please file an issue to discuss or email us contact@kfmdmsolutions.com.

Reporting Bugs
Please ensure to report bugs against an official release. Please try the latest official release before reporting a bug, just in case the latest version has fixed issues that you might be facing.